Rename computation-engine folder to engine
