+++ src/engine/manual/variableExtractor.ts
