+++ src/engine/singular-formula-solver.ts
