+++ src/engine/symbolic-algebra/symbolic-algebra.ts
